move: Component 패키지 이동
